Fix bug in resolver if function is already guessed
If a compile-time function had a macro invocation which generated a
reference the compile-time function had already guessed, the reference
would never be resolved.

They are now resolved like the other invocations, which means the
environment is checked for the status of the reference and guessing
happens accordingly, then and there.