More statement usability improvements
* Added ternary operator (?)
* Added unless
* Added "in" as the scope resolution operator (::)
* Added on-call for turning pathing, in, field, etc. into function
  calls
* Finally added if. It accepts both a single statement (one block only,
essentially the same as (when)) and an else clause. The error will
warn you if you try to do more than 2 blocks, and tell you what to do
to get around it. I also caused scope to automatically be stripped
such that there's no extra braces in every if block
* Fix bug where module filenames could be deallocated or moved even
after tokens are referring to them. Filenames are now properly managed