Added protections for preventing late renames
This ensures that if you rename, you can be sure your new version is
the only one used in the entire environment.

The whole system is pretty janky. I'm going to march forward for now,
and if it ends up being a popular feature, I can try to make it less
terrible.