Fix resolver not handling references in guessed functions
If a reference was added after guessing a function, another
dependency propagation pass is required to make that referenced
definition buildable. Before this change, that additional pass would
only occur if some references were coincidentally resolved.

It now occurs if any new guesses are made. It could be slightly more
efficient to only require propagation if the definitions references
change, but it'd only save one cycle at most, and was a bit obnoxious
implementation-wise.