Added include scanning
I'm not very happy with it, but it is an important feature. If, for
example, you update one of your dependencies, and in that update they
change the size of a struct defined in one of their headers, Cakelisp
will need to invalidate the object files it has cached and recompile
them, because e.g. your sizeof() calls on that struct would be wrong.

The unfortunate part is that this requires scanning hundreds of header
files, and all their includes, every time you want to see if a cached
object can be reused.