Check build commands for changes
This change makes it so changing the build command in any way (new
includes, warning flags, debug/optimization, etc.) will cause a
rebuild of that artifact.

This is really nice when you are iterating on build commands, but can
be troublesome when building multiple different projects in the same
configuration. If the cache isn't being used due to CRC loops, you
probably need to make more configurations.