Fixed header scanning cache problems
This fixed the following problem, which I immediately encountered in
GameLib:
A.cpp includes O.h
B.cpp includes O.h. B.cpp is only included in project B. project B has
A.cpp and B.cpp.

Build project A, which only has A.cpp. A.cpp finds O.h is modified, so
A.cpp is rebuilt.

Build project B, which has A.cpp and B.cpp. B.cpp coincidentally gets
checked first, and finds O.h is modified. B.cpp is built. A.cpp checks
to see whether O.h is modified and gets "true" (all checks being
against the first file to check them, which is B.cpp), so A.cpp is
re-built, unnecessarily.

To make things worse, running build again now causes project A to
re-link, because A.o is marked newer than project A's
executable/library. Having a build do anything after a build and no
modifications is a sin in my book, so I had to fix it.

The solution: the cache now exhaustively determines the maximum
modification time of a header tree, then that time is compared against
each file which includes it. This will be slower for deep header trees
that aren't included in any other build object; otherwise, perf should
be about the same.