Keep delimiter on the context
This fixes the following code:

(defstruct vec3
  x float
  y float
  z float)

;; Expand vector to each of its components
(defmacro vec3-xyz (vec symbol)
  (tokenize-push output
                 (field (token-splice vec) x)
                 (field (token-splice vec) y)
                 (field (token-splice vec) z))
  (return true))

(defun test--math (&return int)
  (var my-vec vec3 (array 1.f 2.f 3.f))
  (printf "%f %f %f\n" (vec3-xyz my-vec))
  (return 0))

On the printf line, the ExpressionList delimiter is lost because macro
expansion uses a different EvaluateGenerateAll_Recursive call. It is
needed on the context so it can be re-evaluated at any time and get
the same result.