Cleared up docs about being Lisp vs. S expressions
I admit that the name "Cakelisp" brings a lot of baggage that makes
what Cakelisp actually is less clear.

The name is very good for SEO, so I'm keeping it. I expect programmers
would call it Lisp if they saw it, simply because the code uses
S-expression syntax and borrows some keywords (defun, defmacro, when,
unless, etc.). A more accurate name would be CakeC, because the actual
mechanics of the language are much closer to C/C++.