Comptime parallel processes, validate CStatements
* Added macro run-process-start-or to make it easier to set up
parallel processes
* Fixed relational operators not properly handling more than two
arguments. Instead of having to decide how to evaluate the statement,
they enforce two args and are wrapped in parens. Not my favorite
option, but it'll do for now
* CStatementOutput now validates for extraneous arguments. This
prevents a bad case of bug where you change an invocation but nothing
happens, because the argument is unused