Added ability to specify hook order of execution
"Priority" isn't the best name, but it's all I could think of for
now. This was important to add for e.g. a pre-build hook which
generates files and a pre-build hook which copies those files. The
generate-files hook must run before the copy-files. It can now easily
be indicated via adding e.g. ":priority-increase 1" to the
generate-files add-compile-time-hook invocation,
or ":priority-decrease 1" to copy-files.