Added first half of tokenize-push rewrite
This rewrite accomplishes many goals by solving all the limitations of
tokenize-push. The new version reuses the tokens already in memory to
output tokens, which gives the following benefits:
* Token sources (line and column numbers) are more accurate
* Macros are unlimited in length (previously, there was a max of 1024
  unspliced characters, which I ended up hitting)
* Strings no longer need extra delimiting, which was a source of
  frustration and bugs
* Performance is better for several reasons: no more tokenization from
  strings, greatly reduced macro compilation time (fewer symbols in
  each macro's generated .cpp file)

It works by saving a pointer to the tokens in the
ObjectDefinition. This pointer is later retrieved by the macro at
runtime via the definition name and tokens CRC. I decided to use a CRC
because it was the only stable identifier I could think of. The
following were considered and rejected as identifiers:
* File/line number. This causes unnecessary recompilation if the file
  the macro definition is in is modified
* Incrementing counter. This doesn't work because macros within
  macros can cause non-sequential incrementation due to unpredictable
  macro resolve/evaluates

While this system is better in every other way, it does require all
macros be evaluated each time cakelisp is run in order to have the
tokens loaded. This may be an issue if incremental compilation is ever
introduced, because now the macro's tokens need to be loaded by some
separate system. The old tokenize-push generated source files were
completely self-contained.

The next half will be executing the push, which will involve
re-interpreting the tokens at macro runtime.