Make it possible to output headers only
The issue was a module which only did:
  (c-import &with-decls "Blah.h")
This broke when imported because the module was empty, so the header
was also not written. Now, the writer will write whatever isn't empty.

While I was at it, I simplified the writer to all be in one for loop
instead of several.