Major new language feature - defer
Added defer generator. This generator will defer outputting until a
scope is exited. This adds some burden on generator writers to be
aware of the need to document their scope enter/exit and explicit
exits.

It was implemented by putting the defer output into a splice, then
having the Writer keep track of scopes and output the splices in the
appropriate locations.

* Speculatively fix double std::move on newStringOutput which seems to
be completely invalid, but didn't seem to break anything.
* Add documentation for path and field